+++ calibrate-angular/src/app/classes/library.ts
@@ -1,3 +1,5 @@
+import { Question } from './question';
+
 export class Library {
 
     libraryId: number;
@@ -11,23 +13,4 @@ enum Status {
     PRIVATE = 0,
     PENDING = 1,
     PUBLIC = 2
-}
-
-export class Question {
-
-    questionId: number;
-    difficulty: number;
-    value: string;
-    libraryId: number;
-    answers: Answers[];
-
-}
-
-class Answers {
-
-    answerId: number;
-    isCorrect: boolean;
-    value: string;
-    questionId: number;
-    
 }
