+++ calibrate-angular/src/app/classes/question.ts
@@ -1,0 +1,18 @@
+export class Question {
+
+    questionId: number;
+    difficulty: number;
+    value: string;
+    libraryId: number;
+    answers: Answers[];
+
+}
+
+class Answers {
+
+    answerId: number;
+    isCorrect: boolean;
+    value: string;
+    questionId: number;
+    isSelected: boolean;
+}
